100_DL_ObjectDetection/requirements.txt
torch
torchvision
opencv-python
Pillow
numpy
pandas
matplotlib
pyyaml
tqdm
scikit-learn
ultralytics>=8.0.196
opencv-python
pandas
pyyaml

